requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
scrapy
selenium
pymongo
spacy
scikit-learn
numpy
pandas
requests
beautifulsoup4
python-dotenv
